+++ lib/src/socket.dart
@@ -1,4 +1,4 @@
-part of dartzmq;
+part of 'zeromq.dart';
 
 /// All types of sockets
 enum SocketType {
@@ -12,15 +12,19 @@ enum SocketType {
   sub,
 
   /// [ZMQ_REQ] = 3
+  /// Synchronous version of [ZMQ_DEALER]
   req,
 
   /// [ZMQ_REP] = 4
+  /// Synchronous version of [ZMQ_ROUTER]
   rep,
 
   /// [ZMQ_DEALER] = 5
+  /// Asynchronous version of [ZMQ_REQ]
   dealer,
 
   /// [ZMQ_ROUTER] = 6
+  /// Asynchronous version of [ZMQ_REP]
   router,
 
   /// [ZMQ_PULL] = 7
@@ -36,19 +40,47 @@ enum SocketType {
   xSub,
 
   /// [ZMQ_STREAM] = 11
-  stream
+  stream,
+
+  /// [ZMQ_SERVER] = 12
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  server,
+
+  /// [ZMQ_CLIENT] = 13
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  client,
+
+  /// [ZMQ_RADIO] = 14
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  radio,
+
+  /// [ZMQ_DISH] = 15
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  dish,
+
+  /// [ZMQ_CHANNEL] = 16
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  channel,
+
+  /// [ZMQ_PEER] = 17
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  peer,
+
+  /// [ZMQ_RAW] = 18
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  raw,
+
+  /// [ZMQ_SCATTER] = 19
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  scatter,
+
+  /// [ZMQ_GATHER] = 20
+  /// Note: This pattern is still in draft state and thus might not be supported by the zeromq library you’re using!
+  gather,
 }
 
-/// ZeroMQ sockets present an abstraction of an asynchronous message queue,
-/// with the exact queuing semantics depending on the socket type in use.
-/// Where conventional sockets transfer streams of bytes or discrete datagrams,
-/// ZeroMQ sockets transfer discrete messages.
-///
-/// ZeroMQ sockets being asynchronous means that the timings of the physical
-/// connection setup and tear down, reconnect and effective delivery are
-/// transparent to the user and organized by ZeroMQ itself. Further, messages
-/// may be queued in the event that a peer is unavailable to receive them.
-class ZSocket {
+/// Base socket
+class ZBaseSocket {
   /// Native socket
   final ZMQSocket _socket;
 
@@ -58,92 +90,78 @@ class ZSocket {
   /// Has this socket been closed?
   bool _closed = false;
 
-  /// StreamController for providing received [ZMessage]s as a stream
-  late final StreamController<ZMessage> _controller;
-
-  /// Stream of received [ZMessage]s
-  Stream<ZMessage> get messages => _controller.stream;
-
-  /// Stream of received [ZFrame]s (expanded [messages] stream)
-  Stream<ZFrame> get frames => messages.expand((element) => element._frames);
-
-  /// Stream of received payloads (expanded [frames] strem)
-  Stream<Uint8List> get payloads => frames.map((e) => e.payload);
-
-  /// Construct a new [ZSocket] with a given underlying ZMQSocket [_socket] and the global ZContext [_context]
-  ZSocket(this._socket, this._context) {
-    _controller = StreamController(onListen: () {
-      _context._listen(this);
-    }, onCancel: () {
-      _context._stopListening(this);
-    });
-  }
+  /// Construct a new [ZBaseSocket] with a given underlying ZMQSocket [_socket] and the global ZContext [_context]
+  ZBaseSocket(this._socket, this._context);
 
   /// Sends the given [data] payload over this socket.
   ///
-  /// The [more] parameter (defaults to false) signals that this is a multi-part
+  /// The [flags] argument is a combination of the flags defined below:
+  ///
+  /// [ZMQ_SNDMORE] signals that this is a multi-part
   /// message. ØMQ ensures atomic delivery of messages: peers shall receive
   /// either all message parts of a message or none at all.
   ///
-  /// The [nowait] parameter (defaults to false) specifies that the operation
+  /// [ZMQ_DONTWAIT] specifies that the operation
   /// should be performed in non-blocking mode. For socket types (DEALER, PUSH)
   /// that block when there are no available peers (or all peers have full
   /// high-water mark). If the message cannot be queued on the socket,
   /// the zmq_send() function shall fail with errno set to EAGAIN.
   ///
   /// Throws [ZeroMQException] on error
-  void send(final List<int> data,
-      {final bool more = false, final bool nowait = false}) {
+  void send(final List<int> data, {final int flags = 0}) {
     _checkNotClosed();
     final ptr = malloc.allocate<Uint8>(data.length);
     ptr.asTypedList(data.length).setAll(0, data);
 
-    final sendParams = more ? ZMQ_SNDMORE : 0 | (nowait ? ZMQ_DONTWAIT : 0);
-    final result =
-        _bindings.zmq_send(_socket, ptr.cast(), data.length, sendParams);
+    final result = _bindings.zmq_send(_socket, ptr.cast(), data.length, flags);
     malloc.free(ptr);
     _checkReturnCode(result, ignore: [EINTR]);
   }
 
   /// Sends the given [string] over this socket
   ///
-  /// The [more] parameter (defaults to false) signals that this is a multi-part
+  /// The [flags] argument is a combination of the flags defined below:
+  ///
+  /// [ZMQ_SNDMORE] signals that this is a multi-part
   /// message. ØMQ ensures atomic delivery of messages: peers shall receive
   /// either all message parts of a message or none at all.
   ///
-  /// The [nowait] parameter (defaults to false) specifies that the operation
+  /// [ZMQ_DONTWAIT] specifies that the operation
   /// should be performed in non-blocking mode. For socket types (DEALER, PUSH)
   /// that block when there are no available peers (or all peers have full
   /// high-water mark). If the message cannot be queued on the socket,
   /// the zmq_send() function shall fail with errno set to EAGAIN.
   ///
   /// Throws [ZeroMQException] on error
-  void sendString(final String string,
-      {final bool more = false, final bool nowait = false}) {
+  void sendString(final String string, {final int flags = 0}) {
     send(
       string.codeUnits,
-      more: more,
-      nowait: nowait,
+      flags: flags,
     );
   }
 
   /// Sends the given [frame] over this socket
   ///
   /// This is a convenience function and is the same as calling
-  /// [send(frame.payload, more: frame.hasMore)]
+  /// [send(frame.payload, flags: frame.hasMore ? ZMQ_SNDMORE : 0)]
   ///
-  /// The [nowait] parameter (defaults to false) specifies that the operation
+  /// The [flags] argument is a combination of the flags defined below:
+  ///
+  /// [ZMQ_SNDMORE] signals that this is a multi-part
+  /// message. ØMQ ensures atomic delivery of messages: peers shall receive
+  /// either all message parts of a message or none at all.
+  ///
+  /// [ZMQ_DONTWAIT] specifies that the operation
   /// should be performed in non-blocking mode. For socket types (DEALER, PUSH)
   /// that block when there are no available peers (or all peers have full
   /// high-water mark). If the message cannot be queued on the socket,
   /// the zmq_send() function shall fail with errno set to EAGAIN.
   ///
   /// Throws [ZeroMQException] on error
-  void sendFrame(final ZFrame frame, {final bool nowait = false}) {
+  void sendFrame(final ZFrame frame, {final int flags = 0}) {
     send(
       frame.payload,
-      more: frame.hasMore,
-      nowait: nowait,
+      flags: flags | (frame.hasMore ? ZMQ_SNDMORE : 0),
     );
   }
 
@@ -152,20 +170,25 @@ class ZSocket {
   /// This is a convenience function.
   /// Note that the individual [ZFrame.hasMore] are ignored
   ///
-  /// The [nowait] parameter (defaults to false) specifies that the operation
+  /// The [flags] argument is a combination of the flags defined below:
+  ///
+  /// [ZMQ_SNDMORE] signals that this is a multi-part
+  /// message. ØMQ ensures atomic delivery of messages: peers shall receive
+  /// either all message parts of a message or none at all.
+  ///
+  /// [ZMQ_DONTWAIT] specifies that the operation
   /// should be performed in non-blocking mode. For socket types (DEALER, PUSH)
   /// that block when there are no available peers (or all peers have full
   /// high-water mark). If the message cannot be queued on the socket,
   /// the zmq_send() function shall fail with errno set to EAGAIN.
   ///
   /// Throws [ZeroMQException] on error
-  void sendMessage(final ZMessage message, {final bool nowait = false}) {
+  void sendMessage(final ZMessage message, {final int flags = 0}) {
     final lastIndex = message.length - 1;
     for (int i = 0; i < message.length; ++i) {
       send(
         message.elementAt(i).payload,
-        more: i < lastIndex ? true : false,
-        nowait: nowait,
+        flags: flags | (i < lastIndex ? ZMQ_SNDMORE : 0),
       );
     }
   }
@@ -206,19 +229,35 @@ class ZSocket {
     if (!_closed) {
       _context._handleSocketClosed(this);
       _bindings.zmq_close(_socket);
-      _controller.close();
       _closed = true;
     }
   }
 
   /// Set a socket [option] to a specific [value]
   ///
-  /// Throws [ZeroMQException] on error
-  void setOption(final int option, final String value) {
-    final ptr = value.toNativeUtf8();
-    final result = _bindings.zmq_setsockopt(
-        _socket, option, ptr.cast<Uint8>(), ptr.length);
-    malloc.free(ptr);
+  /// Throws [ZeroMQException] or [ArgumentError] on error
+  void setOption(final int option, final dynamic value) {
+    final int result;
+    if (value is String) {
+      final ptr = value.toNativeUtf8();
+      result = _bindings.zmq_setsockopt(
+          _socket, option, ptr.cast<Uint8>(), ptr.length);
+      malloc.free(ptr);
+    } else if (value is int) {
+      final ptr = malloc<Int32>();
+      ptr.value = value;
+      result = _bindings.zmq_setsockopt(
+          _socket, option, ptr.cast<Uint8>(), sizeOf<Int32>());
+      malloc.free(ptr);
+    } else if (value is bool) {
+      final ptr = malloc<Int32>();
+      ptr.value = value ? 1 : 0; // 1 for true, 0 for false
+      result = _bindings.zmq_setsockopt(
+          _socket, option, ptr.cast<Uint8>(), sizeOf<Int32>());
+      malloc.free(ptr);
+    } else {
+      throw ArgumentError('Unsupported type: ${value.runtimeType}');
+    }
     _checkReturnCode(result, ignore: [EINTR]);
   }
 
@@ -284,6 +323,46 @@ class ZSocket {
   }
 }
 
+/// ZeroMQ sockets present an abstraction of an asynchronous message queue,
+/// with the exact queuing semantics depending on the socket type in use.
+/// Where conventional sockets transfer streams of bytes or discrete datagrams,
+/// ZeroMQ sockets transfer discrete messages.
+///
+/// ZeroMQ sockets being asynchronous means that the timings of the physical
+/// connection setup and tear down, reconnect and effective delivery are
+/// transparent to the user and organized by ZeroMQ itself. Further, messages
+/// may be queued in the event that a peer is unavailable to receive them.
+class ZSocket extends ZBaseSocket {
+  /// StreamController for providing received [ZMessage]s as a stream
+  late final StreamController<ZMessage> _controller;
+
+  /// Stream of received [ZMessage]s
+  Stream<ZMessage> get messages => _controller.stream;
+
+  /// Stream of received [ZFrame]s (expanded [messages] stream)
+  Stream<ZFrame> get frames => messages.expand((element) => element._frames);
+
+  /// Stream of received payloads (expanded [frames] strem)
+  Stream<Uint8List> get payloads => frames.map((e) => e.payload);
+
+  /// Construct a new [ZSocket] with a given underlying ZMQSocket [_socket] and the global ZContext [_context]
+  ZSocket(super._socket, super._context) {
+    _controller = StreamController(onListen: () {
+      _context._listen(this);
+    }, onCancel: () {
+      _context._stopListening(this);
+    });
+  }
+
+  @override
+  void close() {
+    if (!_closed) {
+      _controller.close();
+    }
+    super.close();
+  }
+}
+
 /// A socket that is monitored by [ZMonitor] to receive events like
 /// [ZEvent.CONNECTED], [ZEvent.DISCONNECTED] as well as other [ZEvent]s
 class MonitoredZSocket extends ZSocket {
@@ -302,5 +381,60 @@ class MonitoredZSocket extends ZSocket {
   void close() {
     _monitor.close();
     super.close();
+  }
+}
+
+/// A socket that only has blocking synchronized functions for receiving
+class ZSyncSocket extends ZBaseSocket {
+  /// Construct a new [ZSyncSocket] with a given underlying ZMQSocket [_socket] and the global ZContext [_context]
+  ZSyncSocket(super._socket, super._context);
+
+  /// Receive a message from the socket, blocking.
+  ///
+  /// The [flags] argument is a combination of the flags defined below:
+  ///
+  /// [ZMQ_SNDMORE] signals that this is a multi-part
+  /// message. ØMQ ensures atomic delivery of messages: peers shall receive
+  /// either all message parts of a message or none at all.
+  ///
+  /// [ZMQ_DONTWAIT] specifies that the operation
+  /// should be performed in non-blocking mode. For socket types (DEALER, PUSH)
+  /// that block when there are no available peers (or all peers have full
+  /// high-water mark). If the message cannot be queued on the socket,
+  /// the zmq_send() function shall fail with errno set to EAGAIN.
+  ///
+  ///
+  /// Useful for REQ/REP sockets where you want to wait for a reply.
+  ///
+  /// Returns the received message
+  /// Throws a [StateError] when called and this socket is closed
+  /// Throws [ZeroMQException] on error
+  ZMessage recv({int flags = 0}) {
+    _checkNotClosed();
+
+    final frame = ZMQBindings.allocateMessage();
+    var rc = _bindings.zmq_msg_init(frame); // rc == 0
+    _checkReturnCode(rc);
+
+    try {
+      ZMessage zMessage = ZMessage();
+      while (true) {
+        rc = _bindings.zmq_msg_recv(frame, _socket, flags);
+        _checkReturnCode(rc);
+        final data = _bindings.zmq_msg_data(frame).cast<Uint8>();
+        final copyOfData = Uint8List.fromList(data.asTypedList(rc));
+
+        final hasMore = _bindings.zmq_msg_more(frame) != 0;
+
+        zMessage.add(ZFrame(copyOfData, hasMore: hasMore));
+        if (!hasMore) {
+          return zMessage;
+        }
+      }
+    } finally {
+      rc = _bindings.zmq_msg_close(frame); // rc == 0
+      malloc.free(frame);
+      _checkReturnCode(rc);
+    }
   }
 }